requirements.txt
# Step 1 — data ingestion
pymongo>=4.6.0
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.26.0
python-dotenv>=1.0.0
pytest>=7.4.0

# Step 2 — features + ML
torch>=2.1.0
scikit-learn>=1.3.0
mlflow>=2.10.0

# Step 3 — API serving
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
cachetools>=5.3.0
httpx>=0.27.0

# Step 4 — MCP server
mcp>=1.0.0
